requirements.txt
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
pymediawiki==0.7.2
requests==2.28.2
soupsieve==2.4
urllib3==1.26.15
